# docs/superpowers/plans/2026-06-17-api-spec.yaml
openapi: "3.0.3"
info:
  title: Intelligent Foreign Trade Financial API
  description: |
    API for the China foreign trade financial software (智能外贸财务系统).
    Features: AI-powered voucher generation, multi-ledger accounting, period closing,
    OEM commission calculation, print & export, audit trail.

    **Auth:** JWT Bearer token in `Authorization` header. CSRF via `X-CSRF-Token`.
    **Multi-ledger:** Pass `X-Ledger-Id` header to scope all operations to a ledger.
  version: "1.0.0"
  contact:
    name: Trad Account Team
servers:
  - url: http://localhost:8004
    description: Local development (8000-8003 blocked by zombie ports on Windows)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VoucherEntrySchema:
      type: object
      required: [account_code, summary, direction, amount]
      properties:
        account_code: { type: string }
        summary: { type: string, description: 摘要 }
        direction: { type: string, enum: ["借", "贷"] }
        amount: { type: number, exclusiveMinimum: 0 }
        currency_code: { type: string, default: CNY }
        original_amount: { type: number, nullable: true }
        exchange_rate: { type: number, default: 1.0000 }
        partner_id: { type: integer, nullable: true }
    VoucherCreate:
      type: object
      required: [voucher_date, entries]
      properties:
        voucher_date: { type: string, format: date }
        voucher_number: { type: string, description: 凭证字号, auto-generated if omitted }
        attachments_count: { type: integer, default: 0 }
        contract_number: { type: string, nullable: true, description: OEM合同号 }
        entries: { type: array, items: { $ref: '#/components/schemas/VoucherEntrySchema' }, minItems: 2 }
    VoucherEntryResponse:
      type: object
      properties:
        id: { type: integer }
        account_id: { type: integer }
        account: { type: object, properties: { id: { type: integer }, code: { type: string }, name: { type: string } } }
        summary: { type: string }
        direction: { type: string }
        amount: { type: number }
        currency_code: { type: string }
        original_amount: { type: number, nullable: true }
        exchange_rate: { type: number }
        partner_id: { type: integer, nullable: true }
        vat_rate: { type: number, nullable: true }
        vat_amount: { type: number, nullable: true }
    VoucherResponse:
      type: object
      properties:
        id: { type: integer }
        ledger_id: { type: integer }
        voucher_number: { type: string }
        voucher_date: { type: string, format: date }
        attachments_count: { type: integer }
        status: { type: string, enum: [DRAFT, APPROVED, POSTED] }
        source_type: { type: string, nullable: true }
        contract_number: { type: string, nullable: true }
        entries: { type: array, items: { $ref: '#/components/schemas/VoucherEntryResponse' } }
...
    VoucherPrintData:
      type: object
      properties:
        voucher: { $ref: '#/components/schemas/VoucherResponse' }
        ledger_name: { type: string }
        company_name: { type: string }
        print_format: { type: string, enum: [A4, A5, wide_24x14] }

    # --- Reports ---
...
security:
  - BearerAuth: []
paths:
...
  /vouchers:
    get:
      tags: [Vouchers]
      summary: List vouchers with pagination & filters
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [DRAFT, APPROVED, POSTED] }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
        - name: keyword
          in: query
          schema: { type: string, description: 摘要/凭证号搜索 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/VoucherResponse' } }, total: { type: integer } } } } } }
    post:
      tags: [Vouchers]
      summary: Create a voucher manually
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VoucherCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/VoucherResponse' } } } }
...
  /vouchers/{id}:
    get:
      tags: [Vouchers]
      summary: Get voucher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/VoucherResponse' } } } }
  /vouchers/{id}/print:
    get:
      tags: [Vouchers]
      summary: Get print-ready voucher data
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: format
          in: query
          schema: { type: string, enum: [A4, A5, wide_24x14], default: A4 }
      responses:
        '200': { description: Print data, content: { application/json: { schema: { $ref: '#/components/schemas/VoucherPrintData' } } } }
  /vouchers/{id}/post:
    post:
      tags: [Vouchers]
      summary: Post a voucher (过账) — changes status from DRAFT/APPROVED to POSTED
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Posted, content: { application/json: { schema: { $ref: '#/components/schemas/VoucherResponse' } } } }
...